Update Queries to filter options with queries
